Test notification implemented. Notification is send when BluetoothActivity is Stopped.
